+++ apiflask/__init__.py
@@ -16,7 +16,10 @@ from .decorators import output
 from .exceptions import abort
 from .exceptions import HTTPError
 from .schemas import Schema
+from .schemas import PaginationSchema
 from .security import HTTPBasicAuth
 from .security import HTTPTokenAuth
+from .utils import get_reason_phrase
+from .utils import pagination_builder
 
 __version__ = '0.6.0dev'